+++ src/utils/insights/chart-format.ts
